refactoring to use type switch, moved to new file
