+++ handlers.go
@@ -152,74 +152,6 @@ func resolveImageURLs(content map[string]interface{}, APIHost string) {
 	}
 }
 
-func removeEmptyMapFields(content map[string]interface{}) {
-	for key, value := range content {
-		if value == nil {
-			delete(content, key)
-			continue
-		}
-
-		str, ok := value.(string)
-		if ok {
-			if str == "" {
-				delete(content, key)
-			}
-			continue
-		}
-
-		subMap, ok := value.(map[string]interface{})
-		if ok {
-			removeEmptyMapFields(subMap)
-			if len(subMap) == 0 {
-				delete(content, key)
-			}
-			continue
-		}
-
-		slice, ok := value.([]interface{})
-		if ok {
-			content[key] = removeEmptySliceValues(slice)
-		}
-	}
-}
-
-func removeEmptySliceValues(slice []interface{}) []interface{} {
-	sliceCopy := []interface{}{}
-
-	for _, value := range slice {
-		if value == nil {
-			continue
-		}
-
-		str, ok := value.(string)
-		if ok {
-			if str != "" {
-				sliceCopy = append(sliceCopy, value)
-			}
-			continue
-		}
-
-		subMap, ok := value.(map[string]interface{})
-		if ok {
-			removeEmptyMapFields(subMap)
-			if len(subMap) != 0 {
-				sliceCopy = append(sliceCopy, value)
-			}
-			continue
-		}
-
-		subSlice, ok := value.([]interface{})
-		if ok {
-			sliceCopy = append(sliceCopy, removeEmptySliceValues(subSlice))
-			continue
-		}
-
-		sliceCopy = append(sliceCopy, value)
-	}
-
-	return sliceCopy
-}
-
 func (handler contentHandler) getContent(ctx context.Context) (ok bool, statusCode int, resp *http.Response) {
 	uuid := ctx.Value(uuidKey).(string)
 	requestURL := fmt.Sprintf("%s%s", handler.serviceConfig.contentSourceURI, uuid)
